+++ plugin/test/net/groboclown/idea/p4ic/v2/server/cache/state/FileUpdateStateListTest.java
@@ -6,6 +6,7 @@ import net.groboclown.idea.p4ic.v2.server.cache.FileUpdateAction;
 import org.junit.Test;
 
 import java.io.File;
+import java.util.Collections;
 
 import static org.hamcrest.core.Is.is;
 import static org.junit.Assert.assertThat;
@@ -31,14 +32,13 @@ public class FileUpdateStateListTest {
     public void testAdd_same() throws Exception {
         FileUpdateStateList list = new FileUpdateStateList();
         P4FileUpdateState s1 = new P4FileUpdateState(
-                new P4ClientFileMapping(null, createFilePath("a.txt")),
+                new P4ClientFileMapping(null, createFilePath("/a.txt")),
                 1, FileUpdateAction.ADD_EDIT_FILE, true);
         list.add(s1);
         assertThat("one add: " + list, list.copy().size(), is(1));
         P4FileUpdateState s2 = new P4FileUpdateState(
-                new P4ClientFileMapping(null, createFilePath("a.txt")),
+                new P4ClientFileMapping(null, createFilePath("/a.txt")),
                 2, FileUpdateAction.DELETE_FILE, true);
-        assertThat(s1.hashCode(), is(s2.hashCode()));
         list.add(s2);
         assertThat("two adds, same file: " + list, list.copy().size(), is(1));
         assertThat("new copy type",
@@ -81,7 +81,7 @@ public class FileUpdateStateListTest {
     public void testRemove_Same_Depot() throws Exception {
         FileUpdateStateList list = new FileUpdateStateList();
         P4FileUpdateState s1 = new P4FileUpdateState(
-                new P4ClientFileMapping("//depot/1/2/3"),
+                new P4ClientFileMapping("//depot/1/2/3", createFilePath("/1/2/3")),
                 1, FileUpdateAction.ADD_EDIT_FILE, true);
         list.add(s1);
         assertThat("one add: " + list, list.copy().size(), is(1));
@@ -110,32 +110,32 @@ public class FileUpdateStateListTest {
     public void testRemove_Equal_LocalDepot() throws Exception {
         FileUpdateStateList list = new FileUpdateStateList();
         P4FileUpdateState s1 = new P4FileUpdateState(
-                new P4ClientFileMapping("//depot/1/2/3", createFilePath("a.txt")),
+                new P4ClientFileMapping("//depot/1/2/3", createFilePath("/a.txt")),
                 1, FileUpdateAction.ADD_EDIT_FILE, true);
         list.add(s1);
         assertThat("one add: " + list, list.copy().size(), is(1));
 
         P4FileUpdateState s2 = new P4FileUpdateState(
-                new P4ClientFileMapping("//depot/1/2/3", createFilePath("a.txt")),
+                new P4ClientFileMapping("//depot/1/2/3", createFilePath("/a.txt")),
                 2, FileUpdateAction.DELETE_FILE, true);
         list.remove(s2);
-        assertThat("one remove: " + list, list.copy().size(), is(0));
+        assertThat("one remove", list.copy(), is(Collections.<P4FileUpdateState>emptySet()));
     }
 
     @Test
     public void testRemove_Equal_Depot() throws Exception {
         FileUpdateStateList list = new FileUpdateStateList();
         P4FileUpdateState s1 = new P4FileUpdateState(
-                new P4ClientFileMapping("//depot/1/2/3"),
+                new P4ClientFileMapping("//depot/1/2/3", createFilePath("/a/b/c")),
                 1, FileUpdateAction.ADD_EDIT_FILE, true);
         list.add(s1);
-        assertThat("one add: " + list, list.copy().size(), is(1));
+        assertThat("one add", list.copy(), is(Collections.singleton(s1)));
 
         P4FileUpdateState s2 = new P4FileUpdateState(
-                new P4ClientFileMapping("//depot/1/2/3"),
+                new P4ClientFileMapping("//depot/1/2/3", createFilePath("/a/b/c")),
                 2, FileUpdateAction.DELETE_FILE, true);
         list.remove(s2);
-        assertThat("one remove: " + list, list.copy().size(), is(0));
+        assertThat("one remove", list.copy(), is(Collections.<P4FileUpdateState>emptySet()));
     }
 
     private FilePath createFilePath(String f) {